Updated server.py and created database
Proposed changes from last meeting on Friday have been implemented the database is also being tracked, and I used some sample images I downloaded to test sending to the server again.
The retrieve function for now is just a placeholder that will have to change to send back an ImageModel row from the db so that we can send back the classification result to the user and visualize the data.

The images I added could be deleted since they are only for testing but that can be done later when we change the structure of the db to now also store the classification result. After that we can delete the data in it and also the images in uploads folder would need to be deleted since they are no longer being referenced in the db.

The TODO's mention what is missing and essentially left to be done.